slightly faster parsing using mmf.ReadFast instead of read(). signed int for positioning since either way i cant go past 2bn tick arrays due to dotnet limitations (why bigarray was implemented), moved stall check back to update() since it wasnt the cause for formerly broken throttling/skipping thankfully. clamp tick to 0 when seeking back. fixed bug on close() that didnt close events cause i set currsynth empty before the switch case

+++ SharpMIDI.Loading/MIDIStorage.cs
@@ -23,20 +23,20 @@ namespace SharpMIDI
     
     public struct Tempo
     {
-        public uint tick;
-        public uint tempo;
+        public int tick;
+        public uint24 tempo;
     }
     
     public struct SysEx
     {
-        public uint tick;
+        public int tick;
         public byte[] message;
     }
     
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     public struct TickGroup
     {
-        public uint tick;
+        public int tick;
         public uint notecount;
         public long offset;
     }
